fix(docs): splitting dynamodb-table to 2 files  (#5261)
fixes #4885 
So we can index this resource
Please let me know if another split is needed

https://www.loom.com/share/bdf686044a1845adbc001d486013e534?sid=89e06465-2bb2-49c8-a7fd-635b4185bc8a

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [x] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.
