+++ docs/docs/04-standard-library/cloud/endpoint.md
@@ -134,3 +134,143 @@ browserSupport: bool;
 - *Type:* bool
 
 Whether the endpoint is supported through browsers. For UI purposes.
+## API Reference <a name="API Reference" id="API Reference"></a>
+
+### Endpoint <a name="Endpoint" id="@winglang/sdk.cloud.Endpoint"></a>
+
+A cloud Endpoint.
+
+#### Initializers <a name="Initializers" id="@winglang/sdk.cloud.Endpoint.Initializer"></a>
+
+```wing
+bring cloud;
+
+new cloud.Endpoint(url: str, props?: EndpointProps);
+```
+
+| **Name** | **Type** | **Description** |
+| --- | --- | --- |
+| <code><a href="#@winglang/sdk.cloud.Endpoint.Initializer.parameter.url">url</a></code> | <code>str</code> | *No description.* |
+| <code><a href="#@winglang/sdk.cloud.Endpoint.Initializer.parameter.props">props</a></code> | <code><a href="#@winglang/sdk.cloud.EndpointProps">EndpointProps</a></code> | *No description.* |
+
+---
+
+##### `url`<sup>Required</sup> <a name="url" id="@winglang/sdk.cloud.Endpoint.Initializer.parameter.url"></a>
+
+- *Type:* str
+
+---
+
+##### `props`<sup>Optional</sup> <a name="props" id="@winglang/sdk.cloud.Endpoint.Initializer.parameter.props"></a>
+
+- *Type:* <a href="#@winglang/sdk.cloud.EndpointProps">EndpointProps</a>
+
+---
+
+
+
+#### Properties <a name="Properties" id="Properties"></a>
+
+| **Name** | **Type** | **Description** |
+| --- | --- | --- |
+| <code><a href="#@winglang/sdk.cloud.Endpoint.property.node">node</a></code> | <code>constructs.Node</code> | The tree node. |
+| <code><a href="#@winglang/sdk.cloud.Endpoint.property.url">url</a></code> | <code>str</code> | The endpoint url. |
+
+---
+
+##### `node`<sup>Required</sup> <a name="node" id="@winglang/sdk.cloud.Endpoint.property.node"></a>
+
+```wing
+node: Node;
+```
+
+- *Type:* constructs.Node
+
+The tree node.
+
+---
+
+##### `url`<sup>Required</sup> <a name="url" id="@winglang/sdk.cloud.Endpoint.property.url"></a>
+
+```wing
+url: str;
+```
+
+- *Type:* str
+
+The endpoint url.
+
+---
+
+
+
+## Structs <a name="Structs" id="Structs"></a>
+
+### EndpointProps <a name="EndpointProps" id="@winglang/sdk.cloud.EndpointProps"></a>
+
+Options for `Endpoint`.
+
+#### Initializer <a name="Initializer" id="@winglang/sdk.cloud.EndpointProps.Initializer"></a>
+
+```wing
+bring cloud;
+
+let EndpointProps = cloud.EndpointProps{ ... };
+```
+
+#### Properties <a name="Properties" id="Properties"></a>
+
+| **Name** | **Type** | **Description** |
+| --- | --- | --- |
+| <code><a href="#@winglang/sdk.cloud.EndpointProps.property.browserSupport">browserSupport</a></code> | <code>bool</code> | Whether the endpoint is supported through browsers. |
+| <code><a href="#@winglang/sdk.cloud.EndpointProps.property.label">label</a></code> | <code>str</code> | The endpoint's label. |
+
+---
+
+##### `browserSupport`<sup>Optional</sup> <a name="browserSupport" id="@winglang/sdk.cloud.EndpointProps.property.browserSupport"></a>
+
+```wing
+browserSupport: bool;
+```
+
+- *Type:* bool
+- *Default:* undefined
+
+Whether the endpoint is supported through browsers.
+
+For UI purposes.
+
+---
+
+##### `label`<sup>Optional</sup> <a name="label" id="@winglang/sdk.cloud.EndpointProps.property.label"></a>
+
+```wing
+label: str;
+```
+
+- *Type:* str
+- *Default:* undefined
+
+The endpoint's label.
+
+For UI purposes.
+
+---
+
+*Example*
+
+```wing
+"My Dashboard"
+```
+
+
+## Protocols <a name="Protocols" id="Protocols"></a>
+
+### IEndpointClient <a name="IEndpointClient" id="@winglang/sdk.cloud.IEndpointClient"></a>
+
+- *Implemented By:* <a href="#@winglang/sdk.cloud.IEndpointClient">IEndpointClient</a>
+
+Inflight interface for `Endpoint`.
+
+
+
