+++ docs/docs/04-standard-library/ex/dynamodb-table/_category_.yml
@@ -1,0 +1,3 @@
+label: DynamodbTable
+collapsible: true
+collapsed: true
