+++ docs/docs/04-standard-library/std/README.md
@@ -1,0 +1,3 @@
+# Wing Globals
+
+This directory contains types available in the global scope of Wing.
