+++ libs/wingsdk/.projen/tasks.json
@@ -121,12 +121,6 @@
       "description": "Generate API.md from .jsii manifest",
       "steps": [
         {
-          "exec": "cp -r src/cloud/*.md ../../docs/docs/04-standard-library/cloud/"
-        },
-        {
-          "exec": "cp -r src/ex/*.md ../../docs/docs/04-standard-library/ex/"
-        },
-        {
           "exec": "jsii-docgen -o API.md -l wing --submodule expect"
         },
         {
@@ -235,6 +229,9 @@
           "exec": "cat API.md >> ../../docs/docs/04-standard-library/util/api-reference.md"
         },
         {
+          "exec": "cp -r ./src/cloud/*.md ../../docs/docs/04-standard-library/cloud/"
+        },
+        {
           "exec": "jsii-docgen -o API.md -l wing --submodule cloud/api"
         },
         {
@@ -257,6 +254,12 @@
         },
         {
           "exec": "cat API.md >> ../../docs/docs/04-standard-library/cloud/domain.md"
+        },
+        {
+          "exec": "jsii-docgen -o API.md -l wing --submodule cloud/endpoint"
+        },
+        {
+          "exec": "cat API.md >> ../../docs/docs/04-standard-library/cloud/endpoint.md"
         },
         {
           "exec": "jsii-docgen -o API.md -l wing --submodule cloud/function"
@@ -313,10 +316,7 @@
           "exec": "cat API.md >> ../../docs/docs/04-standard-library/cloud/workload.md"
         },
         {
-          "exec": "jsii-docgen -o API.md -l wing --submodule ex/dynamodb-table"
-        },
-        {
-          "exec": "cat API.md >> ../../docs/docs/04-standard-library/ex/dynamodb-table.md"
+          "exec": "cp -r ./src/ex/*.md ../../docs/docs/04-standard-library/ex/"
         },
         {
           "exec": "jsii-docgen -o API.md -l wing --submodule ex/react-app"
@@ -335,6 +335,24 @@
         },
         {
           "exec": "cat API.md >> ../../docs/docs/04-standard-library/ex/table.md"
+        },
+        {
+          "exec": "cp -r ./src/ex/dynamodb-table/*.md ../../docs/docs/04-standard-library/ex/dynamodb-table/"
+        },
+        {
+          "exec": "jsii-docgen -o API.md -l wing --submodule ex/dynamodb-table/interfaces"
+        },
+        {
+          "exec": "cat API.md >> ../../docs/docs/04-standard-library/ex/dynamodb-table/interfaces.md"
+        },
+        {
+          "exec": "jsii-docgen -o API.md -l wing --submodule ex/dynamodb-table/table"
+        },
+        {
+          "exec": "cat API.md >> ../../docs/docs/04-standard-library/ex/dynamodb-table/table.md"
+        },
+        {
+          "exec": "cp -r ./src/std/*.md ../../docs/docs/04-standard-library/std/"
         },
         {
           "exec": "echo \"---\ntitle: array\nid: array\n---\n\n\" > ../../docs/docs/04-standard-library/std/array.md"
