+++ libs/wingsdk/src/ex/dynamodb-table/index.ts
@@ -1,0 +1,2 @@
+export * from "./interfaces";
+export * from "./table";
