+++ libs/wingsdk/src/ex/dynamodb-table/interfaces.md
@@ -1,0 +1,7 @@
+---
+title: Dynamodb Table interfaces
+id: dynamodb-table-interfaces
+description: A set of interfaces used with dynamoDb client.
+keywords: [dynamoDB, NoSQL, store, DB, database, cache, table]
+sidebar_position: 1
+---
