+++ libs/wingsdk/src/ex/dynamodb-table/table.md
@@ -1,6 +1,6 @@
 ---
 title: Dynamodb Table
-id: dynamodb-table
+id: dynamodb-table-table
 description: A cloud Dynamodb table.
 keywords: [dynamoDB, NoSQL, store, DB, database, cache, table]
 sidebar_position: 1
